Introduced OperandStack, OperandQueue, and OperandList
Went through code and comments to unify the use of "queue"
and "list" to match the new naming:

An instruction takes an operand list is input.
Future operand lists reside on the operand queue.
Each function has an operand queue.
The operand stack holds all operand queues for current and calling
functions.